moved tests from ASTRuleTests to RulesTests
