--- Source/SwiftLintFrameworkTests/ASTRuleTests.swift
@@ -1,36 +1,0 @@
-//
-//  ASTRuleTests.swift
-//  SwiftLint
-//
-//  Created by JP Simard on 5/28/15.
-//  Copyright (c) 2015 Realm. All rights reserved.
-//
-
-import SwiftLintFramework
-import XCTest
-
-class ASTRuleTests: XCTestCase {
-    func testTypeBodyLength() {
-        verifyRule(TypeBodyLengthRule.description)
-    }
-
-    func testTypeName() {
-        verifyRule(TypeNameRule.description)
-    }
-
-    func testVariableName() {
-        verifyRule(VariableNameRule.description)
-    }
-
-    func testNesting() {
-        verifyRule(NestingRule.description)
-    }
-
-    func testControlStatement() {
-        verifyRule(ControlStatementRule.description)
-    }
-
-    func testCyclomaticComplexity() {
-        verifyRule(CyclomaticComplexityRule.description)
-    }
-}
